fix linting/tests
