+++ .eslintignore
@@ -2,3 +2,4 @@ node_modules
 coverage
 build
 webpack.*.js
+*.test.js
