--- babel.config.json
@@ -1,29 +1,0 @@
-{
-    "presets": [
-        [
-            "@babel/preset-env",
-            {
-                "targets": "> 1%, not dead"
-            }
-        ],
-        "@babel/preset-typescript",
-        "react-app"
-    ],
-    "plugins": [
-        "@babel/plugin-proposal-class-properties",
-        "react-hot-loader/babel",
-        ["react-css-modules", {
-            "filetypes": {
-                ".scss": { "syntax": "postcss-scss" }
-            },
-            "attributeNames": {
-                "activeStyleName": "activeClassName",
-                "contentStyleName": "contentClassName",
-                "headerStyleName": "headerClassName"
-            },
-            "webpackHotModuleReloading": true,
-            "generateScopedName": "[name]__[local]--[hash:base64:10]",
-            "autoResolveMultipleImports": true
-        }]
-    ]
-}
