+++ src/tapis-app/__tests/Sidebar.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import configureStore from 'redux-mock-store';
 import '@testing-library/jest-dom/extend-expect';
 import renderComponent from './utils/testing';
-import Sidebar from 'tapis-app/Sidebar';
+import Sidebar from 'tapis-app/src/Sidebar';
 import tapisReduxStore from 'fixtures/tapis-redux.fixture';
 
 const mockStore = configureStore();
