+++ src/tapis-app/__tests/utils/testing.js
@@ -2,13 +2,11 @@ import React from 'react';
 import { BrowserRouter, Router } from 'react-router-dom';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
-/* eslint-disable */
-import TapisProvider from 'tapis-hooks/provider';
-
+import TapisProvider from 'tapis-hooks/src/provider';
 export default function renderComponent(component, store, history) {
   if (history) {
     return render(
-      <TapisProvider baseUrl="https://tacc.tapis.io">
+      <TapisProvider basePath="tapis.test">
         <Provider store={store}>
           <Router history={history}>{component}</Router>
         </Provider>
@@ -16,7 +14,7 @@ export default function renderComponent(component, store, history) {
     );
   }
   return render(
-    <TapisProvider baseUrl="https://tacc.tapis.io">
+    <TapisProvider basePath="tapis.test">
       <Provider store={store}>
         <BrowserRouter>{component}</BrowserRouter>
       </Provider>
