+++ src/tapis-ui/_tests/_common/AppIcon.test.js
