+++ src/tapis-ui/_tests/_common/Icon.test.js
