+++ src/tapis-ui/_tests/_common/Paginator.test.js
@@ -4,7 +4,7 @@ import Paginator from 'tapis-ui/src/_common/Paginator';
 
 describe('Paginator', () => {
   it('renders pages', () => {
-    const { getAllByText } = render(<Paginator pages={20} current={11} />);
+    const { getAllByText } = render(<Paginator pages={20} current={11} callback={() => {}} />);
     expect(getAllByText('1').length).toEqual(1);
     expect(getAllByText('9').length).toEqual(1);
     expect(getAllByText('10').length).toEqual(1);
