+++ src/tapis-ui/_tests/components/jobs/JobLauncher.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import configureStore from 'redux-mock-store';
 import '@testing-library/jest-dom/extend-expect';
 import renderComponent from '../../utils/testing';
-import { JobLauncher } from 'tapis-ui/components/jobs';
+import { JobLauncher } from 'tapis-ui/src/components/jobs';
 import { Jobs } from '@tapis/tapis-typescript';
 import tapisReduxStore from 'fixtures/tapis-redux.fixture';
