+++ src/tapis-ui/_tests/components/jobs/JobsListing.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import configureStore from 'redux-mock-store';
 import '@testing-library/jest-dom/extend-expect';
 import renderComponent from '../../utils/testing';
-import { JobsListing } from 'tapis-ui/components/jobs';
+import { JobsListing } from 'tapis-ui/src/components/jobs';
 import tapisReduxStore from 'fixtures/tapis-redux.fixture';
 
 const mockStore = configureStore();
