+++ src/tapis-ui/_tests/components/systems/SystemList.test.tsx
@@ -2,16 +2,16 @@ import React from 'react';
 import configureStore from 'redux-mock-store';
 import '@testing-library/jest-dom/extend-expect';
 import renderComponent from '../../utils/testing';
-import { AppsListing } from 'tapis-ui/components/apps';
+import { SystemList } from 'tapis-ui/src/components/systems';
 import tapisReduxStore from 'fixtures/tapis-redux.fixture';
 
 const mockStore = configureStore();
 
-describe('Apps', () => {
-  it('renders AppsListing component', () => {
+describe('SystemList', () => {
+  it('renders SystemList component', () => {
     const store = mockStore(tapisReduxStore);
 
-    const { getAllByText } = renderComponent(<AppsListing />, store);
-    expect(getAllByText(/SleepSeconds/).length).toEqual(1);
+    const { getAllByText } = renderComponent(<SystemList />, store);
+    expect(getAllByText(/testuser8-e2e/).length).toEqual(1);
   });
 });
