--- webpack.prod.js
@@ -1,98 +1,0 @@
-const { merge } = require('webpack-merge');
-const { CleanWebpackPlugin } = require('clean-webpack-plugin');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const HtmlWebpackHarddiskPlugin = require('html-webpack-harddisk-plugin');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const CompressionPlugin = require('compression-webpack-plugin');
-const TerserJSPlugin = require('terser-webpack-plugin');
-const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
-const webpack = require('webpack');
-const path = require('path');
-const common = require('./webpack.common.js');
-
-module.exports = merge(common, {
-  mode: 'production',
-  output: {
-    path: path.resolve(__dirname, 'build'),
-    publicPath: '/tapis-ui/',
-    filename: '[name].[hash].bundle.js',
-    chunkFilename: '[name].[hash].bundle.js',
-  },
-  module: {
-    rules: [
-      {
-        test: /^(.)*\.module\.(css|sass|scss)$/i,
-        use: [
-          {
-            loader: MiniCssExtractPlugin.loader,
-          },
-          {
-            loader: 'css-loader',
-            options: {
-              sourceMap: true,
-              modules: {
-                localIdentName: '[name]__[local]--[hash:base64:10]',
-              },
-              localsConvention: 'camelCase',
-            },
-          },
-          {
-            loader: 'sass-loader',
-            options: {
-              sourceMap: true,
-            },
-          },
-        ],
-      },
-      {
-        test: /^((?!\.module).)*\.(css|sass|scss)$/i,
-        use: [
-          {
-            loader: MiniCssExtractPlugin.loader,
-          },
-          {
-            loader: 'css-loader',
-            options: {
-              sourceMap: true,
-            },
-          },
-          {
-            loader: 'sass-loader',
-            options: {
-              sourceMap: true,
-            },
-          },
-        ],
-      }
-    ],
-  },
-  plugins: [
-    new webpack.ProvidePlugin({
-      process: 'process/browser',
-    }),
-    new webpack.EnvironmentPlugin({
-      TAPIS_TENANT_URL: 'https://tacc.tapis.io'
-    }),
-    new CleanWebpackPlugin(),
-    new HtmlWebpackPlugin({
-      template: path.resolve(__dirname, 'src/tapis-app/src', 'index.html'),
-      inject: true,
-    }),
-    new MiniCssExtractPlugin({
-      filename: '[name].[hash].css',
-      chunkFilename: '[id].css',
-    }),
-    new CompressionPlugin({
-      // filename: '[path].br[query]',
-      // algorithm: 'brotliCompress',
-      // // test: /\.(js|css|html|svg|png|jpg)$/,
-      // compressionOptions: { level: 11 },
-      // threshold: 10240,
-      // minRatio: 0.8
-    }),
-    new HtmlWebpackHarddiskPlugin(),
-  ],
-  optimization: {
-    minimizer: [new TerserJSPlugin({}), new OptimizeCSSAssetsPlugin({})],
-  },
-});
